src/catalog/api/artists/response.py
from rest_framework import serializers

class ArtistCreateResponse(serializers.Serializer):
    id = serializers.IntegerField(read_only=True)
    name = serializers.CharField()
    created_at = serializers.DateTimeField()