pyproject.toml
[project]
name = "test-qortex"
version = "0.1.0"
description = ""

requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false